code split for API network call
